+++ Communication/Packets/Incoming/Groups/GetGroupInfoEvent.cs
@@ -1,4 +1,5 @@
 using Plus.Communication.Packets.Outgoing.Groups;
+using Plus.HabboHotel.Cache;
 using Plus.HabboHotel.GameClients;
 using Plus.HabboHotel.Groups;
 using Plus.HabboHotel.Rooms;
@@ -9,11 +10,13 @@ internal class GetGroupInfoEvent : IPacketEvent
 {
     private readonly IGroupManager _groupManager;
     private readonly IRoomFactory _roomFactory;
+    private readonly ICacheManager _cacheManager;
 
-    public GetGroupInfoEvent(IGroupManager groupManager, IRoomFactory roomFactory)
+    public GetGroupInfoEvent(IGroupManager groupManager, IRoomFactory roomFactory, ICacheManager cacheManager)
     {
         _groupManager = groupManager;
         _roomFactory = roomFactory;
+        _cacheManager = cacheManager;
     }
 
     public Task Parse(GameClient session, IIncomingPacket packet)
@@ -22,7 +25,7 @@ internal class GetGroupInfoEvent : IPacketEvent
         var newWindow = packet.ReadBool();
         if (!_groupManager.TryGetGroup(groupId, out var group))
             return Task.CompletedTask;
-        session.Send(new GroupInfoComposer(group, session, _roomFactory, newWindow));
+        session.Send(new GroupInfoComposer(group, session, _roomFactory, _cacheManager, newWindow));
         return Task.CompletedTask;
     }
 }