+++ Communication/Packets/Incoming/Rooms/Furni/GetGroupFurniSettingsEvent.cs
@@ -1,4 +1,5 @@
 using Plus.Communication.Packets.Outgoing.Groups;
+using Plus.HabboHotel.Cache;
 using Plus.HabboHotel.GameClients;
 using Plus.HabboHotel.Groups;
 using Plus.HabboHotel.Items;
@@ -10,11 +11,13 @@ internal class GetGroupFurniSettingsEvent : IPacketEvent
 {
     private readonly IGroupManager _groupManager;
     private readonly IRoomFactory _roomFactory;
+    private readonly ICacheManager _cacheManager;
 
-    public GetGroupFurniSettingsEvent(IGroupManager groupManager, IRoomFactory roomFactory)
+    public GetGroupFurniSettingsEvent(IGroupManager groupManager, IRoomFactory roomFactory, ICacheManager cacheManager)
     {
         _groupManager = groupManager;
         _roomFactory = roomFactory;
+        _cacheManager = cacheManager;
     }
 
     public Task Parse(GameClient session, IIncomingPacket packet)
@@ -35,7 +38,7 @@ internal class GetGroupFurniSettingsEvent : IPacketEvent
             return Task.CompletedTask;
 
         session.Send(new GroupFurniSettingsComposer(group, itemId, habbo.Id));
-        session.Send(new GroupInfoComposer(group, session, _roomFactory));
+        session.Send(new GroupInfoComposer(group, session, _roomFactory, _cacheManager));
         return Task.CompletedTask;
     }
 }